group game, with fixes

+++ app/models/motile.rb
@@ -19,13 +19,15 @@ class Motile < ActiveRecord::Base
         rations = Ration.joins(:game_user).where(position_id: self.position_id, game_users: {game_id: self.game.id})
         rations.each do |ration|
           ingredient = Ingredient.where(ration_id: ration.id).first
-          name = ingredient.ingredient_cat.name
-          messages.push({
-            title:'Motile Collision',
-            text:'A ration belonging to '+ration.game_user.user.name+' was hit by a motile peice and lost a <span class="label '+name+'">'+name+'</span>',
-            type:'warning', time: 6
-          })
-          ingredient.destroy
+          if ingredient
+            name = ingredient.ingredient_cat.name
+            messages.push({
+              title:'Motile Collision',
+              text:'A ration belonging to '+ration.game_user.user.name+' was hit by a motile peice and lost a <span class="label '+name+'">'+name+'</span>',
+              type:'warning', time: 6
+            })
+            ingredient.destroy
+          end
         end
       else
         # how did this happen?
